requirements.txt
face_recognition==1.3.0
opencv-python==4.10.0.84
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.66.0
flask>=3.0.0
dlib>=19.24.0
